+++ src/main/java/xyz/openmodloader/registry/Delegate.java
@@ -1,0 +1,11 @@
+package xyz.openmodloader.registry;
+
+import java.lang.annotation.ElementType;
+import java.lang.annotation.Retention;
+import java.lang.annotation.RetentionPolicy;
+import java.lang.annotation.Target;
+
+@Retention(value = RetentionPolicy.RUNTIME)
+@Target(ElementType.FIELD)
+public @interface Delegate {
+}
