+++ src/test/resources/META-INF/MANIFEST.MF
@@ -7,3 +7,5 @@ Author: OML
 URL: http://example.com/
 Update-URL: https://gist.githubusercontent.com/anonymous/bf49748c02fc854be2cc312110b91ebc/raw/ad770f1c717a4d22eb6e89984fec8cdcffe560ea/2016-06-26_20-53-08.txt
 Logo: logo.png
+Client-Delegate: xyz.openmodloader.test.OMLClientDelegate
+Server-Delegate: xyz.openmodloader.test.OMLServerDelegate
